refactor: split core modules
